Use correct data type when calling NumericAddressedBackend::read.
There are two versions of NumericAddressedBackend::read, and neither of
them expects values of type size_t for the first two arguments, so
calling it with such values results in an ambiguity which version should
be used.

By making sure that we call it with arguments of type uint64_t, we can
resolve this ambiguity.

+++ device_backends/NumericAddressedBackend/include/NumericAddressedLowLevelTransferElement.h
@@ -163,10 +163,10 @@ namespace ChimeraTK {
     boost::shared_ptr<NumericAddressedBackend> _dev;
 
     /** start address w.r.t. the PCIe bar */
-    size_t _bar;
+    uint64_t _bar;
 
     /** start address w.r.t. the PCIe bar */
-    size_t _startAddress;
+    uint64_t _startAddress;
 
     /** number of bytes to access */
     size_t _numberOfBytes;
